Physics Slight Fix
Added built Box2D stuff, linked Box2D, split my box test into an implementation in the main.cpp as well as in separate physics.cpp and physics.h files. Though the code for both are essentially the same, the implementation in the main.cpp file works successfuly. Use of the physics.h and physics.cpp file results in the LNK2019 error.